requirements.txt
Flask~=3.1.2
Flask-PyMongo
Flask-Admin
PyJWT~=2.10.1
werkzeug~=3.1.3
python-dotenv~=1.2.1
gunicorn
WTForms~=3.2.1
pymongo~=4.15.4
email-validator